+++ CHANGELOG.md
@@ -18,6 +18,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Included quick links in the search results of the assistant
 - Added a skeleton loader to the changelog page
 - Extended the content of the _Self-Hosting_ section by information about additional data providers on the Frequently Asked Questions (FAQ) page
+- Enhanced the create or update activity dialog to display the current holdings as default options
 
 ### Changed